tests: Prioritize device config with inet IP.
Previously RDMATestCase randomly chose which device config to work with.
With this new change it randomly selects ports with inet IP configured,
in case there are any, instead of selecting from all ports found.
After this change CMTestCase can rely on the device selection that was
made by its parent, RDMATestCase.
